requirements.txt
attrs==21.4.0
black==21.12b0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.10
click==8.0.3
cryptography==36.0.1
Flask==2.0.2
greenlet==1.1.2
gunicorn==20.1.0
idna==3.3
importlib-metadata==4.10.1
iniconfig==1.1.1
itsdangerous==2.0.1
Jinja2==3.0.3
markdown2==2.4.2
MarkupSafe==2.0.1
mypy-extensions==0.4.3
packaging==21.3
pathspec==0.9.0
platformdirs==2.4.1
pluggy==1.0.0
psutil==5.9.0
py==1.11.0
pycparser==2.21
pydantic==1.9.0
Pygments==2.11.2
PyJWT==2.3.0
PyMySQL==1.0.2
pyparsing==3.0.7
pytest==6.2.5
requests==2.27.1
SQLAlchemy==1.4.31
sqlalchemy2-stubs==0.0.2a19
sqlmodel==0.0.6
toml==0.10.2
tomli==1.2.3
typing_extensions==4.0.1
urllib3==1.26.8
Werkzeug==2.0.2
zipp==3.7.0
